Add `force` option
When force is true, upon completion of all tests, test failures will not
cause the whole task to fail, in either the case of a negative
assertion, or when there were no assertions at all. This can be set to
true when you always want other tasks in the queue to be executed.

This is effectively implements --force at a task-specific level, where
it is common to want to continue even when tests fail.

Certain PhantomJS error conditions have also been changed to raise test
failures rather than exiting the task immediately.