Update to Puppeteer 5
Some internal breaking changes, but nothing that affects this project
afaik. https://github.com/puppeteer/puppeteer/releases/tag/v5.0.0

Main reason for updating is so that 5.3.0 becomes in-range,
which introduced support for PUPPETEER_DOWNLOAD_PATH, which can
be set to a cached directory (e.g. for CI to save/restore, or to
persist locally in a way compatible with npm-ci)

Ref https://github.com/puppeteer/puppeteer/pull/6014