Restore page timeout handling and `fail.timeout` event
Follows b35e7ce1b4 and https://github.com/gruntjs/grunt-contrib-qunit/pull/147,
which converted the plugin from PhantomJS to Headless Chromium.

While `options.timeout` is still used for the navigation, it was no
longer used to stop tasks that have crashed, broken or are timing out
for other reasons at runtime.

The approach of using an interval in the bridge is based on the
original grunt-lib-phantomjs code [1]. I modified it to use a more
graceful setTimeout recursion instead, and changed the "end" logic
from a direct `exit()` call (which is not possible in chrome bridge),
to simply not recursing further.

The exit handling is already in place in tasks/qunit.js, where an
`on("fail.*")` listener exists (though seemingly not utilised until
now), which includes an indirect call to `browser.close()`.

[1] https://github.com/gruntjs/grunt-lib-phantomjs/blob/v1.1.0/phantomjs/main.js#L43-L51

Fixes https://github.com/gruntjs/grunt-contrib-qunit/issues/178.