Migrate reporting logic to `QUnit.on()` events API
== Why ==

* Higher level API, exists for this purpose, thus resulting in simpler code
  that has to know less about QUnit internals.

* Solves a number of problems, including two issues that would make CI on this
  repo fail if we updated its devDependencies to the latest qunit as-is.

== Change ==

Remove all manual tracking of `QUnit.log()` and custom interpreting of results.
Simply listen to `QUnit.on()` for "testStart", "testEnd" and "runEnd", and
then report their counts, details, and boolean status as-is.

The main observable change is that the printed success line is now one line
instead of two:

Before:
> 2 tests completed with 0 failed, 0 skipped, and 0 todo.
> 2 assertions (in 9ms), passed: 2, failed: 0

After:
> 2 tests completed in 9ms, with 0 failed, 0 skipped, and 0 todo.

This is the same text as displayed in the browser.

== Details ==

The task previously tracked individually logged assertions, decided if it's a
failure, and filtered out expected failures in a "todo" test from actually
unexpected failures.

This resulted in the odd situation where `QUnit.done()` is given a count but
the task did not use it to decide whether the grunt task is a success, because
this would not account for "todo" tests, and thus in
https://github.com/gruntjs/grunt-contrib-qunit/pull/137 the logic grew even
further. In QUnit 2.17 the reporting of global failures from window.onerror
changed and the logic here was now stale and no longer reporting it correctly.

Also, the qunit_noglobals.html case was wrongly counted as a failing test
because the logic checked `failed === 0 && passed === total` based on raw
assertion count, which again, does not account for "todo" tests.
The `qunit_noglobals.html` is actually a passing test when opened in the
browser, but the local Gruntfile logic treated it as a failure.

Both of these two problems are naturally fixed by this patch, and so the next
commit that updates QUnit in this repo's CI to the latest will continue to pass
instead of having these two failures.