docs: Document minimum QUnit version and regenerate README
As of grunt-contrib-qunit 6, the plugin uses `QUnit.on()`. This API
was introduced in 2017 early on in the QUnit 2.x series, 2.2.0 to be
exact.

If a project is upgrading from the PhantomJS-era all the way to the
latest grunt-contrib-qunit with Headless Chromium, then it may run
into an error like `Error: TypeError: QUnit.on is not a function` if
they haven't yet also bumped the included qunit version e.g. if they
still used QUnit 1.x.

Document the minimum QUnit version as such.