Release 8.0.0: upgrade to Puppeteer 21, require Node 16+
From grunt-contrib-connect 4.0.0:
> https://github.com/gruntjs/grunt-contrib-connect/pull/282
> * Fix deprecation warning about `opn@6`.

From puppeteer 20:
> https://github.com/puppeteer/puppeteer/releases/tag/puppeteer-v20.0.0
> * switch from Chromium to "Chrome for Testing".

From puppeteer 20.9.0:
> https://github.com/puppeteer/puppeteer/releases/tag/puppeteer-core-v20.9.0
> * roll to Chrome 115

From puppeteer 21.1.0:
> https://github.com/puppeteer/puppeteer/releases/tag/puppeteer-core-v21.1.0
> * roll to Chrome 116

General "Chrome for Testing" announcement at [1].
Based on reading the linked design document [2] [3], I believe
"Chrome for Testing" continues to be a build of open-source Chromium.
It is not compiled from or with proprietary Google Chrome components.
Rather, it is a Chromium build that is 1) compiled feature flags and
defaults settings as similar to Google Chrome as possible, and 2)
built and published from the same Chromium revisions and at the same
cadence as Chrome releases. This means, that anytime a Google Chrome
release happens for Stable or Beta channel, there is also a
Chrome for Testing release.

Having said that, the design document does acknowledge that a previous
proposal (nicknamed "Chrome-", meaning start with Chrome and subtract
unwanted features) did exist. This was rejected in favour of "Chromium+"
which works as above with the Chrome-like app icon injected from their
build system from outside the Chromium source tree.

[1] https://developer.chrome.com/blog/chrome-for-testing/
[2] https://docs.google.com/document/d/1XJvxyqAQjhPfJ0rX84PjfXXb5sBx3m8DXzMxR0ipQNs/edit
[3] http://web.archive.org/web/20230503104427/https://docs.google.com/document/d/1XJvxyqAQjhPfJ0rX84PjfXXb5sBx3m8DXzMxR0ipQNs/edit

From puppeteer 19.11:
> https://github.com/puppeteer/puppeteer/releases/tag/puppeteer-core-v19.11.0
> * add warning for `headless: true`

There is no more pretending that Chrome Headless wasn't actually
"Chrome but headless" (like Firefox Headless), but a separate browser
implementation [4]. "Fake it until you make it". Starting in Chrome 112,
there is a "new" headless mode, which is now claimed to actually be
"Chrome, but headless". I propose that in grunt-contrib-qunit 8.0,
with the switch to "Chromium for Testing", we also switch to this
mode at the same time.

If we don't, Puppeteer prints the following warning on every CI build:

> Puppeteer old Headless deprecation warning:
> In the near feature `headless: true` will default to the new Headless mode
> Consider opting in early by passing `headless: "new"` to `puppeteer.launch()`

[4] https://developer.chrome.com/articles/new-headless/