Remove most QUnit callback forwarding in favour of QUnit events
Especially, remove the deprecated details parameter to `QUnit.done()`
as this will be undefined in QUnit 3.0 and thus would cause
grunt-contrib-qunit to crash due to reading `obj.failed` on undefined.

https://qunitjs.com/api/callbacks/QUnit.done/

Most of these are obsoleted by QUnit.on, which are already forwarded
to Grunt's event system as `'qunit.on.*'` and can be used instead.

https://qunitjs.com/api/callbacks/QUnit.on/

To help in the common case, and to ease migration/debugging, leave
forwarding of begin/done events in place but without parameters.