require Node 18+
This is in prep for Puppeteer 22, which requires Node 18+. [1]

grunt-contrib-internal does not yet have a version that provisions
CI with Node 18+ matrix, which has delayed/complicated changes in this
repo before as well so I'm going to opt-out from the contrib-ci
template enforcement for now. This seems to be common in half a dozen
other first-party grunt repos as well. [2]

[1]: https://github.com/puppeteer/puppeteer/blob/puppeteer-v22.10.0/packages/puppeteer-core/CHANGELOG.md#2200-2024-02-05
[2]: https://github.com/search?q=skipIfExists%20org%3Agruntjs&type=code