upgrade to Puppeteer 22
Puppeteer 22 drops support for Node 16 and requires Node 18+ [1].

The main change for us is how `headless` is configured. [2] The new
headless mode is now enabled by default in Puppeteer, corresponding
to `headless: true` instead of `headless: 'new'`.

We already opted-in to this before in grunt-contrib-qunit so it is not
relevant for grunt-contrib-qunit release notes.

This also includes "roll to Chrome 125", from puppeteer 22.10.0 [3].

[1]: https://github.com/puppeteer/puppeteer/releases/tag/puppeteer-core-v22.0.0
[2]: https://github.com/puppeteer/puppeteer/issues/11815
[3]: https://github.com/puppeteer/puppeteer/blob/puppeteer-v22.10.0/packages/puppeteer/CHANGELOG.md