Remove support for delaying qunit.js via AMD
The current check in bridge.js is problematic because if a project
includes RequireJS anywhere, for any purpose, the bridge crashes
unless RequireJS is also in charge of loading QUnit.

QUnit 3.0 will remove support for its built-in AMD export
(ref https://github.com/qunitjs/qunit/issues/1729).

It remains supported to use AMD/RequireJS in your HTML file, and to
use it to load source code and/or test files. However, it is no
longer supported to delay the loading of qunit.js itself via AMD.

In order to make that use possible in QUnit 3.0, the bridge.js must
stop assuming that presence of define.amd implies using it to load
QUnit itself.

Example at https://qunitjs.com/api/config/autostart/#loading-with-requirejs