Include output from `QUnit.on('error')`
The two are not likely to overlap, but this could in theory report
the same error twice. In practice, given the bridge kicks in fairly
late, uncaught errors will happen before our QUnit listener is
active, and late errors are likely exclusive to `QUnit.on('error')`.

The "No tests" message is changing in QUnit 3.0 from a "fake" test,
to an error event, hence this is in preparation for that.

Note that the `runEnd` event, which is use to decide on the exit code,
already reflects whether there were errors, so this doesn't effect
the (default) build status, hence planning to release in a minor
version.